requirements.txt
streamlit>=1.29
pillow
streamlit_image_coordinates
